使用rxbus+retrofit+okio完善下载图片出现容易出现oom问题

+++ app/src/main/java/com/example/yuxuehai/wallpager/ui/fragment/PhotoesDetailFragment.java
@@ -207,7 +207,7 @@ public class PhotoesDetailFragment extends MvpBaseFragment<PhotoView,
                 .alpha(1)
                 .setDuration(300)
                 .start();
-        mProgressBar.setProgress(event.getProgress());
+        mProgressBar.setProgress((int) event.getProgress());
         if (event.getProgress() == 100) {
             mProgressBar.animate()
                     .alpha(0)
